requirements.txt
googlesearch-python
beautifulsoup4

# DeSOTA Packgs
pyyaml
requests